requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
dash==2.10.2
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
fonttools==4.40.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
psycopg2-binary==2.9.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
seaborn==0.12.2
six==1.16.0
SQLAlchemy==2.0.16
tenacity==8.2.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.2.3
zipp==3.15.0
